backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
opencv-python==4.10.0.84
mediapipe==0.10.14
openai-whisper==20231117  # Correct PyPI name
moviepy==1.0.3
pydub==0.25.1
transformers==4.44.0
torch==2.4.0
groq==0.9.0
pillow==10.4.0
numpy==2.0.1
python-dotenv==1.0.1
pyaudio==0.2.14
#pip install opencv-python mediapipe